+++ fastai/data_block.py
@@ -13,8 +13,9 @@ def _maybe_squeeze(arr): return (arr if is1d(arr) else np.squeeze(arr))
 
 def _get_files(parent, p, f, extensions):
     p = Path(p)#.relative_to(parent)
+    low_extensions = [e.lower() for e in extensions] if extensions is not None else None
     res = [p/o for o in f if not o.startswith('.')
-           and (extensions is None or f'.{o.split(".")[-1].lower()}' in extensions)]
+           and (extensions is None or f'.{o.split(".")[-1].lower()}' in low_extensions)]
     return res
 
 def get_files(path:PathOrStr, extensions:Collection[str]=None, recurse:bool=False,
@@ -289,7 +290,7 @@ class CategoryProcessor(PreProcessor):
 
     def generate_classes(self, items):
         "Generate classes from `items` by taking the sorted unique values."
-        return uniqueify(items)
+        return uniqueify(items, sort=True)
 
     def process_one(self,item):
         if isinstance(item, EmptyLabel): return item
@@ -496,6 +497,7 @@ class LabelLists(ItemLists):
         if getattr(self, 'normalize', False):#In case a normalization was serialized
             norm = self.normalize
             data.normalize((norm['mean'], norm['std']), do_x=norm['do_x'], do_y=norm['do_y'])
+        data.label_list = self
         return data
 
     def add_test(self, items:Iterator, label:Any=None):