+++ fastai/text/learner.py
@@ -17,7 +17,7 @@ __all__ = ['RNNLearner', 'LanguageLearner', 'convert_weights', 'decode_spec_toke
 _model_meta = {AWD_LSTM: {'hid_name':'emb_sz', 'url':URLs.WT103_1,
                           'config_lm':awd_lstm_lm_config, 'split_lm': awd_lstm_lm_split,
                           'config_clas':awd_lstm_clas_config, 'split_clas': awd_lstm_clas_split},
-               Transformer: {'hid_name':'d_model', 
+               Transformer: {'hid_name':'d_model', 'url':URLs.OPENAI_TRANSFORMER,
                              'config_lm':tfmer_lm_config, 'split_lm': tfmer_lm_split,
                              'config_clas':tfmer_clas_config, 'split_clas': tfmer_clas_split},
                TransformerXL: {'hid_name':'d_model', 
@@ -26,18 +26,19 @@ _model_meta = {AWD_LSTM: {'hid_name':'emb_sz', 'url':URLs.WT103_1,
 
 def convert_weights(wgts:Weights, stoi_wgts:Dict[str,int], itos_new:Collection[str]) -> Weights:
     "Convert the model `wgts` to go with a new vocabulary."
-    dec_bias, enc_wgts = wgts['1.decoder.bias'], wgts['0.encoder.weight']
-    bias_m, wgts_m = dec_bias.mean(0), enc_wgts.mean(0)
+    dec_bias, enc_wgts = wgts.get('1.decoder.bias', None), wgts['0.encoder.weight']
+    wgts_m = enc_wgts.mean(0)
+    if dec_bias is not None: bias_m = dec_bias.mean(0)
     new_w = enc_wgts.new_zeros((len(itos_new),enc_wgts.size(1))).zero_()
-    new_b = dec_bias.new_zeros((len(itos_new),)).zero_()
+    if dec_bias is not None: new_b = dec_bias.new_zeros((len(itos_new),)).zero_()
     for i,w in enumerate(itos_new):
         r = stoi_wgts[w] if w in stoi_wgts else -1
         new_w[i] = enc_wgts[r] if r>=0 else wgts_m
-        new_b[i] = dec_bias[r] if r>=0 else bias_m
+        if dec_bias is not None: new_b[i] = dec_bias[r] if r>=0 else bias_m
     wgts['0.encoder.weight'] = new_w
-    wgts['0.encoder_dp.emb.weight'] = new_w.clone()
+    if '0.encoder_dp.emb.weight' in wgts: wgts['0.encoder_dp.emb.weight'] = new_w.clone()
     wgts['1.decoder.weight'] = new_w.clone()
-    wgts['1.decoder.bias'] = new_b
+    if dec_bias is not None: wgts['1.decoder.bias'] = new_b
     return wgts
 
 class RNNLearner(Learner):
@@ -124,9 +125,9 @@ class LanguageLearner(RNNLearner):
             new_idx.append(idx)
             xb = xb.new_tensor([idx])[None]
         return text + sep + sep.join(decoder(self.data.vocab.textify(new_idx, sep=None)))
-    
-    def beam_search(self, text:str, n_words:int, top_k:int=10, beam_sz:int=1000, temperature:float=1., sep:str=' ',
-                    decoder=decode_spec_tokens):
+
+    def beam_search(self, text:str, n_words:int, no_unk:bool=True, top_k:int=10, beam_sz:int=1000, temperature:float=1.,
+                    sep:str=' ', decoder=decode_spec_tokens):
         "Return the `n_words` that come after `text` using beam search."
         ds = self.data.single_dl.dataset
         self.model.reset()
@@ -134,12 +135,13 @@ class LanguageLearner(RNNLearner):
         nodes = None
         xb = xb.repeat(top_k, 1)
         nodes = xb.clone()
-        scores = xb.new_ones(1).float()
+        scores = xb.new_zeros(1).float()
         with torch.no_grad():
             for k in progress_bar(range(n_words), leave=False):
                 out = F.log_softmax(self.model(xb)[0][:,-1], dim=-1)
+                if no_unk: out[:,self.data.vocab.stoi[UNK]] = -float('Inf')
                 values, indices = out.topk(top_k, dim=-1)
-                scores = (-values * scores[:,None]).view(-1)
+                scores = (-values + scores[:,None]).view(-1)
                 indices_idx = torch.arange(0,nodes.size(0))[:,None].expand(nodes.size(0), top_k).contiguous().view(-1)
                 sort_idx = scores.argsort()[:beam_sz]
                 scores = scores[sort_idx]
@@ -149,8 +151,8 @@ class LanguageLearner(RNNLearner):
                 self.model[0].select_hidden(indices_idx[sort_idx])
                 xb = nodes[:,-1][:,None]
         if temperature != 1.: scores.div_(temperature)
-        node_idx = torch.multinomial(1-scores, 1).item()
-        return text + sep + sep.join(decoder(self.data.vocab.textify([i.item() for i in nodes[node_idx]], sep=None)))
+        node_idx = torch.multinomial(torch.exp(-scores), 1).item()
+        return sep.join(decoder(self.data.vocab.textify([i.item() for i in nodes[node_idx][1:] ], sep=None)))
 
     def show_results(self, ds_type=DatasetType.Valid, rows:int=5, max_len:int=20):
         from IPython.display import display, HTML